+++ zetaclient/chains/zrepo/client.go
@@ -3,8 +3,6 @@ package zrepo
 import (
 	"context"
 
-	cosmosmath "cosmossdk.io/math"
-	upgrade "cosmossdk.io/x/upgrade/types"
 	cometbft "github.com/cometbft/cometbft/types"
 	eth "github.com/ethereum/go-ethereum/common"
 	"github.com/zeta-chain/go-tss/blame"
@@ -15,21 +13,54 @@ import (
 	fungible "github.com/zeta-chain/node/x/fungible/types"
 	observer "github.com/zeta-chain/node/x/observer/types"
 	keys "github.com/zeta-chain/node/zetaclient/keys/interfaces"
-	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
-
-var _ ZetacoreClient = &zetacore.Client{}
 
 type ChainID = int64
 type Nonce = uint64
 
-// ZetacoreWriter contains the functions that mutate ZetaChain state.
-type ZetacoreWriter interface {
+type ZetacoreClient interface {
+	zetacoreReaderClient
+	zetacoreWriterClient
+}
+
+// zetacoreReaderClient contains the functions that do not mutate ZetaChain state.
+type zetacoreReaderClient interface {
+	Chain() chains.Chain
+
+	NewBlockSubscriber(context.Context) (chan cometbft.EventDataNewBlock, error)
+
+	HasVoted(_ context.Context, ballotIndex string, voterAddress string) (bool, error)
+
+	ListPendingCCTX(context.Context, chains.Chain) ([]*crosschain.CrossChainTx, uint64, error)
+
+	GetCctxByNonce(context.Context, ChainID, Nonce) (*crosschain.CrossChainTx, error)
+
+	GetCctxByHash(context.Context, string) (*crosschain.CrossChainTx, error)
+
+	GetOutboundTracker(context.Context, ChainID, Nonce) (*crosschain.OutboundTracker, error)
+
+	GetOutboundTrackers(context.Context, ChainID) ([]crosschain.OutboundTracker, error)
+
+	GetInboundTrackersForChain(context.Context, ChainID) ([]crosschain.InboundTracker, error)
+
+	GetKeys() keys.ObserverKeys
+
+	GetForeignCoinsFromAsset(context.Context, ChainID, eth.Address) (fungible.ForeignCoins, error)
+
+	GetPendingNoncesByChain(context.Context, ChainID) (observer.PendingNonces, error)
+
+	GetBallotByID(context.Context, string) (*observer.QueryBallotByIdentifierResponse, error)
+
+	GetBTCTSSAddress(context.Context, ChainID) (string, error)
+}
+
+// zetacoreWriterClient contains the functions that mutate ZetaChain state.
+type zetacoreWriterClient interface {
 	PostVoteGasPrice(_ context.Context,
 		_ chains.Chain,
 		gasPrice uint64,
 		priorityFee uint64,
-		blockNum uint64,
+		block uint64,
 	) (string, error)
 
 	PostVoteTSS(_ context.Context,
@@ -64,79 +95,54 @@ type ZetacoreWriter interface {
 	) (string, error)
 }
 
-// ZetacoreClientRepo contains the functions used by ZetaRepo.
-type ZetacoreClientRepo interface {
-	ZetacoreWriter
+// ------------------------------------------------------------------------------------------------
+// dry-mode
+// ------------------------------------------------------------------------------------------------
 
-	Chain() chains.Chain
+const unreachableMsg = "called an unreachable dryZetacoreClient function"
 
-	GetKeys() keys.ObserverKeys
+// dryZetacoreClient is a dry-wrapper for the zetacore client.
+// It overrides mutating functions from the underlying client that panic when called.
+type dryZetacoreClient struct {
+	zetacoreReaderClient
 
-	ListPendingCCTX(context.Context, chains.Chain) ([]*crosschain.CrossChainTx, uint64, error)
-
-	GetForeignCoinsFromAsset(context.Context, ChainID, eth.Address) (fungible.ForeignCoins, error)
-
-	GetPendingNoncesByChain(context.Context, ChainID) (observer.PendingNonces, error)
-
-	GetCctxByNonce(context.Context, ChainID, Nonce) (*crosschain.CrossChainTx, error)
-
-	GetCctxByHash(context.Context, string) (*crosschain.CrossChainTx, error)
-
-	GetBallotByID(context.Context, string) (*observer.QueryBallotByIdentifierResponse, error)
-
-	GetOutboundTracker(context.Context, ChainID, Nonce) (*crosschain.OutboundTracker, error)
-
-	GetOutboundTrackers(context.Context, ChainID) ([]crosschain.OutboundTracker, error)
-
-	GetInboundTrackersForChain(context.Context, ChainID) ([]crosschain.InboundTracker, error)
-
-	NewBlockSubscriber(context.Context) (chan cometbft.EventDataNewBlock, error)
-
-	GetBTCTSSAddress(context.Context, ChainID) (string, error)
-
-	HasVoted(context.Context, string, string) (bool, error)
+	// writerClient is deliberately not embedded so the compiler can ensure that all mutating
+	// methods are explicitly overridden.
+	writerClient zetacoreWriterClient
 }
 
-// ZetacoreClient is the client interface that interacts with zetacore.
-//
-// TODO: this should be moved elsewhere, since it is not used by ZetaRepo.
-// See: https://github.com/zeta-chain/node/issues/4300
-//
-//go:generate mockery --name ZetacoreClient --filename zetacore_client.go --case underscore --output ../../testutils/mocks
-type ZetacoreClient interface {
-	ZetacoreClientRepo
+func newDryZetacoreClient(client ZetacoreClient) *dryZetacoreClient {
+	return &dryZetacoreClient{zetacoreReaderClient: client, writerClient: client}
+}
 
-	GetSupportedChains(context.Context) ([]chains.Chain, error)
+func (*dryZetacoreClient) PostVoteGasPrice(context.Context, chains.Chain, uint64, uint64, uint64,
+) (string, error) {
+	panic(unreachableMsg)
+}
 
-	GetAdditionalChains(context.Context) ([]chains.Chain, error)
+func (*dryZetacoreClient) PostVoteTSS(context.Context, string, int64, chains.ReceiveStatus,
+) (string, error) {
+	panic(unreachableMsg)
+}
 
-	GetChainParams(context.Context) ([]*observer.ChainParams, error)
+func (*dryZetacoreClient) PostVoteBlameData(context.Context, *blame.Blame, ChainID, string,
+) (string, error) {
+	panic(unreachableMsg)
+}
 
-	GetKeyGen(context.Context) (observer.Keygen, error)
+func (*dryZetacoreClient) PostVoteOutbound(context.Context, uint64, uint64,
+	*crosschain.MsgVoteOutbound,
+) (string, string, error) {
+	panic(unreachableMsg)
+}
 
-	GetTSS(context.Context) (observer.TSS, error)
-	GetTSSHistory(context.Context) ([]observer.TSS, error)
+func (*dryZetacoreClient) PostVoteInbound(context.Context, uint64, uint64,
+	*crosschain.MsgVoteInbound, chan<- zetaerrors.ErrTxMonitor,
+) (string, string, error) {
+	panic(unreachableMsg)
+}
 
-	GetBlockHeight(context.Context) (int64, error)
-
-	ListPendingCCTXWithinRateLimit(context.Context,
-	) (*crosschain.QueryListPendingCctxWithinRateLimitResponse, error)
-
-	GetRateLimiterInput(_ context.Context,
-		window int64,
-	) (*crosschain.QueryRateLimiterInputResponse, error)
-
-	GetCrosschainFlags(context.Context) (observer.CrosschainFlags, error)
-	GetRateLimiterFlags(context.Context) (crosschain.RateLimiterFlags, error)
-	GetOperationalFlags(context.Context) (observer.OperationalFlags, error)
-
-	GetObserverList(context.Context) ([]string, error)
-
-	GetZetaHotKeyBalance(context.Context) (cosmosmath.Int, error)
-
-	GetUpgradePlan(context.Context) (*upgrade.Plan, error)
-
-	GetBaseGasPrice(context.Context) (int64, error)
-
-	GetNumberOfUnconfirmedTxs(context.Context) (int, error)
+func (*dryZetacoreClient) PostOutboundTracker(context.Context, ChainID, Nonce, string,
+) (string, error) {
+	panic(unreachableMsg)
 }